requirements.txt
flask>=2.3.3
tflite-runtime==2.14.0 ; platform_system != "Windows"
tensorflow-cpu==2.20.0 ; platform_system == "Windows"
opencv-python>=4.9.0.80
numpy>=1.24.0
gunicorn>=21.2.0
